ping replacements little more eagerly
previously pings only happened when there were bad entries
and we had replacements. but not when the bucket was partially filled
and we had replacements.
This was in an attempt to let buckets drain so they could be merged,
but that logic is not sound, we should just merge non-home buckets
when necessary